refactor: createImageUploader 팩토리 함수를 packages/hooks로 분리하여 중복 로직 제거
