+++ packages/hooks/index.ts
@@ -2,3 +2,4 @@ export * from './useDebouncedValue';
 export * from './useInfiniteScroll';
 export * from './useAbortableLifeCycle';
 export * from './useKeyboardOffset';
+export * from './createImageUploader';
