+++ packages/hooks/package.json
@@ -7,6 +7,9 @@
     "build": "tsc -p tsconfig.json",
     "lint": "eslint ."
   },
+  "dependencies": {
+    "@surf/utils": "workspace:*"
+  },
   "peerDependencies": {
     "react": "^19"
   },
